test: Fix last test, md5 (needed to do long -> int for uint32).

+++ testsuite/tests/md5/md5.h
@@ -29,8 +29,10 @@
  ************************************************************************
  */
 
+#include <stdint.h>
+
 /* typedef a 32 bit type */
-typedef unsigned long int UINT4;
+typedef uint32_t UINT4;
 
 /* Data structure for MD5 (Message Digest) computation */
 typedef struct {
